my_agent/requirements.txt
langgraph
langchain_community
python-dotenv 
psycopg[binary]
qdrant-client
langchain
langchain_openai
structlog
pydantic
python-dotenv
qdrant-client  
langchain-openai
langchain-google-genai
langchain-anthropic
infisicalsdk
logging_config
pandas
openpyxl
matplotlib